requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.29.7
azure-eventhub==5.11.5
azure-eventhub-checkpointstoreblob==1.1.4
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-identity==1.15.0
azure-search-documents==11.4.0b11
azure-storage-blob==12.19.0
backoff==2.2.1
banal==1.0.6
beautifulsoup4==4.12.3
blinker==1.6.3
blis==0.7.11
bs4==0.0.2
cachetools==5.3.2
case-conversion==2.1.0
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clipboard==0.0.4
cloudpathlib==0.16.0
colorama==0.4.6
colour==0.1.5
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cryptography==42.0.1
cursor==1.3.5
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
dataset==1.6.2
debugpy==1.8.0
decorator==4.4.2
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
Flask==2.3.3
fonttools==4.47.2
frozenlist==1.4.1
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-auth==2.23.3
google-search-results==2.4.1
googleapis-common-protos==1.61.0
greenlet==3.0.3
h11==0.14.0
html-text==0.5.2
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.1
infi.systray==0.1.12
inflect==7.0.0
ipykernel==6.29.0
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==24.2.0
kiwisolver==1.4.5
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langcodes==3.3.0
langsmith==0.0.84
lib==4.0.0
lxml==5.1.0
Mako==1.2.4
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
maybe-else==0.2.1
mbstrdecoder==1.1.3
more-itertools==10.1.0
moviepy==1.0.3
msal==1.26.0
msal-extensions==1.1.0
msoffcrypto-tool==5.3.1
msrest==0.7.1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
O365==2.0.33
oauthlib==3.2.2
office365==0.3.15
Office365-REST-Python-Client==2.5.5
olefile==0.47
openai==1.10.0
opencensus==0.11.3
opencensus-context==0.1.3
opencensus-ext-azure==1.1.11
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
pandas==2.1.2
parsedatetime==2.6
parso==0.8.3
pathmagic==0.3.14
pillow==10.2.0
pipreqs==0.4.13
platformdirs==4.1.0
portalocker==2.8.2
preshed==3.0.9
prettierfier==1.0.3
proglog==0.1.10
prompt-toolkit==3.0.43
promptflow==0.1.0b8
promptflow-tools==0.1.0b10
protobuf==4.24.4
psutil==5.9.8
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydash==6.0.2
pydub==0.25.1
Pygments==2.17.2
pyinstrument==4.6.2
pyiotools==0.3.18
PyJWT==2.8.0
pymiscutils==0.3.14
pyparsing==3.1.1
pypdf==4.0.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pysubtypes==0.3.18
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
random2==1.0.1
readchar==4.0.5
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
srsly==2.4.8
stack-data==0.6.3
strictyaml==1.7.3
stringcase==1.2.0
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.2
tiktoken==0.5.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typepy==1.3.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.0
waitress==2.1.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
XlsxWriter==3.1.9
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
